Update AddCategory, add CategoryExists, CategoryIdByName, AddPostCategories
AddCategory now takes a string as parameter instead of a models.PostCategory.
CategoryExists is needed as the user is able to provide new categories, so I need to check if they're already in the DB.
The front is sending me the names of the categories, so I need to get their IDs to connect the post and the categories.
AddPostCategories is a wrapper for adding multiple categories, as the front is sending me a slice of category names.